backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.5.0
argon2-cffi==23.1.0
PyJWT==2.8.0
stripe==7.11.0
requests==2.31.0
urllib3==2.1.0
reportlab==4.0.8
openai==1.7.2

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Celery & Redis
celery==5.3.6
redis==5.0.1
flower==2.0.1

# ChromaDB for Vector Search (Week 7-8)
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu==1.7.4
numpy==1.26.3

# Azure OpenAI
openai==1.7.2
tiktoken==0.5.2

# Security Scanning Tools (Python wrappers)
python-nmap==0.7.1
shodan==1.31.0

# Report Generation
reportlab==4.0.8
weasyprint==60.2
jinja2==3.1.3
markdown==3.5.2

# AWS S3 for file storage
boto3==1.34.34

# Environment & Config
python-dotenv==1.0.1

# Utilities
aiohttp==3.9.3
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2

# Security
cryptography==42.0.2
pyotp==2.9.0
qrcode==7.4.2

# Docker Integration (Week 5-6)
docker==7.0.0

# Monitoring & Logging
prometheus-client==0.19.0

# Email Delivery (Week 9-10)
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.3

# Production Deployment
gunicorn==21.2.0
sentry-sdk[fastapi]==1.40.0
